move formatter to utils/line/formatGetBorrower, remove some ts ignores, format rates
