+++ src/client/routes/Portfolio/index.tsx
@@ -14,8 +14,8 @@ import {
 } from '@store';
 import { SummaryCard, ViewContainer, NoWalletCard, SliderCard, LineDetailsDisplay } from '@components/app';
 import { SpinnerLoading, Text } from '@components/common';
-import { halfWidthCss, isValidAddress, normalize } from '@utils';
-import { AggregatedCreditLine, CreditLinePage, CreditPosition } from '@src/core/types';
+import { halfWidthCss, isValidAddress, normalize, formatGetBorrowerQuery } from '@utils';
+import { CreditLinePage, CreditPosition } from '@src/core/types';
 
 const StyledViewContainer = styled(ViewContainer)`
   display: grid;
@@ -151,54 +151,9 @@ export const Portfolio = () => {
   }, [borrowerPositions]);
 
   useEffect(() => {
-    let aggregate: CreditLinePage = {
-      borrower: borrowerAddress,
-      principal: '0', //done
-      deposit: '0', //done
-      highestApy: ['0', '0', '0'], //todo
-      positions: [], //done
-      escrow: {
-        id: '',
-        cratio: '',
-        minCRatio: '',
-        collateralValue: '',
-        //@ts-ignore
-        deposits: [],
-      }, //todo
-      spigot: {
-        id: '',
-        tokenRevenue: {},
-      }, //todo
-
-      //CreditLinePage data
-      interest: '0', //done
-      totalInterestRepaid: '0', //to review
-      collateralEvents: [], //todo
-      creditEvents: [], //done
-    };
+    let aggregate;
     if (data) {
-      data.map((data: any, i) => {
-        data.positions.map((position: CreditPosition, i: number) => {
-          const normalizedPrincipal = normalize('amount', position.principal, position.token.decimals);
-          const normalizedDeposit = normalize('amount', position.deposit, position.token.decimals);
-          const normalizedInterest = normalize('amount', position.interestAccrued, position.token.decimals);
-          const normalizedInterestRepaid = normalize('amount', position.interestRepaid, position.token.decimals);
-          const totalPrincipal = Number(aggregate.principal) + Number(normalizedPrincipal);
-          const totalDeposit = Number(aggregate.deposit) + Number(normalizedDeposit);
-          const totalInterest = Number(aggregate.interest) + Number(normalizedInterest);
-          const totalInterestRepaid = Number(aggregate.totalInterestRepaid) + Number(normalizedInterestRepaid);
-          //@ts-ignore
-          aggregate.positions.push(position!);
-          aggregate.principal = `${totalPrincipal}`;
-          aggregate.deposit = `${totalDeposit}`;
-          aggregate.interest = `${totalInterest}`;
-          aggregate.totalInterestRepaid = `${totalInterestRepaid}`;
-        });
-        data.events.map((event: any, i: number) => {
-          aggregate.creditEvents.push(event);
-        });
-      });
-      console.log(aggregate);
+      aggregate = formatGetBorrowerQuery(data, borrowerAddress);
     }
     setAggregatedCreditLine(aggregate);
   }, [data]);
